// StaffTravel/StaffTravel/src/app/models/basic-info.ts
//export class BasicInfo {
//    empNumber: number;
//    firstName: string;
//    lastName: string;
//    seniorityDate: Date;
//    howLong: string;
//    isFullTime: boolean;
//    howManyPassesYouHave: number;
//    yearlyUsed: number;
//    bonusUsed: number;
//    lastMinuteUsed: number;
//    bonusEarned: number
//}

export class BasicInfo {
    empnumber: number;
    firstname: string;
    lastname: string;
    senioritydate: Date;
    HowLong: string;
    IsFullTime: boolean;
    HowManyPassesYouHave: number;
    YearlyUsed: number;
    BonusUsed: number;
    LastMinuteUsed: number;
    BonusEarned: number
    YealyPending: number;
    BonusPending: number;
    LastPending: number;
}